fix compilation

+++ system.h
@@ -13,7 +13,7 @@
 #define SYSTEM_H
 
 long TempsCenti();
-int inter();
+long inter();
 void MakePath(char* path, const char* filename);
 
 #endif
